requirements.txt
llama-extract
python-dotenv
streamlit
pandas
yfinance
plotly
python-dotenv>=0.19.0
streamlit>=1.22.0
pandas>=1.3.0
numpy>=1.21.0 